Only use PieceDragFeedback with the editor
