+++ lib/chessground.dart
@@ -15,7 +15,6 @@ export 'src/premove.dart';
 export 'src/utils.dart';
 export 'src/widgets/board.dart';
 export 'src/widgets/board_editor.dart';
-export 'src/widgets/drag.dart';
 export 'src/widgets/highlight.dart';
 export 'src/widgets/piece.dart';
 export 'src/widgets/background.dart';
