+++ lib/src/widgets/board_editor.dart
@@ -3,9 +3,8 @@ import 'package:flutter/widgets.dart';
 
 import '../board_editor_settings.dart';
 import '../models.dart';
-import 'board.dart';
-import 'drag.dart';
 import '../fen.dart';
+import 'board.dart';
 import 'piece.dart';
 import 'positioned_square.dart';
 
@@ -113,16 +112,12 @@ class _BoardEditorState extends State<ChessBoardEditor> {
                   Draggable(
                     hitTestBehavior: HitTestBehavior.translucent,
                     data: piece,
-                    feedback: BoardDragFeedback(
+                    feedback: PieceDragFeedback(
                       squareSize: widget.squareSize,
                       scale: widget.settings.dragFeedbackScale,
                       offset: widget.settings.dragFeedbackOffset,
-                      child: PieceWidget(
-                        piece: piece,
-                        size: widget.squareSize *
-                            widget.settings.dragFeedbackScale,
-                        pieceAssets: widget.settings.pieceAssets,
-                      ),
+                      piece: piece,
+                      pieceAssets: widget.settings.pieceAssets,
                     ),
                     childWhenDragging: const SizedBox.shrink(),
                     onDragStarted: () => draggedPieceOrigin = squareId,
@@ -199,5 +194,47 @@ class _BoardEditorState extends State<ChessBoardEditor> {
     if (squareId != null) {
       widget.onTouchedSquare?.call(squareId);
     }
+  }
+}
+
+/// The [Piece] to show under the pointer when a drag is under way.
+///
+/// You can use this to drag pieces onto a [ChessBoardEditor] with the same appearance as when the pieces on the board are dragged.
+class PieceDragFeedback extends StatelessWidget {
+  const PieceDragFeedback({
+    super.key,
+    required this.piece,
+    required this.squareSize,
+    required this.pieceAssets,
+    this.scale = 2.0,
+    this.offset = const Offset(0.0, -1.0),
+  });
+
+  /// The piece that is being dragged.
+  final Piece piece;
+
+  /// Size of a square on the board.
+  final double squareSize;
+
+  /// Scale factor for the feedback widget.
+  final double scale;
+
+  /// Offset the feedback widget from the pointer position.
+  final Offset offset;
+
+  /// Piece set
+  final PieceAssets pieceAssets;
+
+  @override
+  Widget build(BuildContext context) {
+    final feedbackSize = squareSize * scale;
+    return Transform.translate(
+      offset: (offset - const Offset(0.5, 0.5)) * feedbackSize / 2,
+      child: PieceWidget(
+        piece: piece,
+        size: feedbackSize,
+        pieceAssets: pieceAssets,
+      ),
+    );
   }
 }
